fixed a nav bar issue in the shelter dashboard. consolidated changes for various endpoints to look consistent
