+++ server/application/app.py
@@ -5,9 +5,8 @@ from flask import Flask, send_from_directory
 from dotenv import load_dotenv
 
 from application.rest import work_shift
-from application.rest import service_commitment
-
-from application.rest.shelter.shelter import shelter_blueprint
+from application.rest.service_commitment import service_commitment_bp
+from application.rest.shelter import shelter_blueprint
 from application.rest.service_shifts import service_shift_bp
 
 
@@ -28,7 +27,7 @@ def create_app(config_name = "development"):
 
     app.register_blueprint(work_shift.blueprint)
     app.register_blueprint(shelter_blueprint)
-    app.register_blueprint(service_commitment.blueprint)
+    app.register_blueprint(service_commitment_bp)
     app.register_blueprint(service_shift_bp)
     load_dotenv()  # Load environment variables from the .env file
